some small fixups - adds API to expose memory handle
